restructure notebooks into A / B / T lines
Line A (NOAA) and line B (DS0N) now run the same eight steps through the same
functions, so 03A and 03B read side by side and differ only in loader and
thresholds. Line T is for analyses that have not earned a step number yet.

  01A_download_data          registry moved to config; one-off SHARP query kept local
  02A_data_processing        pipeline now in src/processing.py; the `regions[-1:]`
                             slice replaced by a real loop with SKIP_EXISTING
  03A_data_analysis          8 steps x 5 regions, params from config.REGION_PARAMS;
                             gains spectra, detrending and the mu geometry check
  03B_data_analysis          new; the same 8 steps over DS00-DS09, replacing 03T's
                             twelve copy-pasted blocks with retyped thresholds
  04_data_comparison         gains 03A's cross-region tail (summary, 24 h fits,
                             Wilson-depression scatter); DS0N sections cleaned
  01T_coefficient_recon.     was 01B; parked feature, repointed at post_processing
  03T_analysis_sandbox       new; seeded with the questions still open
  archive/                   01_download_data, 03_data_analisis, 03T's DS0N batch

Fixes found while wiring it up:

- the raw-cube flags for NOAA 11106 lived in the loader parameters, so they would
  have flowed into 04 and turned its fitted amplitude into an uncalibrated number
  (raw v_SDO is ~3 km/s and diurnal, right on the period being measured). Now an
  explicit opt-in comparison step, available for every region.
- DS0N cubes were read as float64 though they are float32 on disk (BITPIX -32),
  doubling memory for no precision. Peak RSS for DS00 drops 9.5 -> 5.5 GB and 03B
  now completes where it was previously OOM-killed. Means accumulate in float64
  regardless, so DS0N metrics stay bit-identical; NOAA metrics change by at most
  6e-5 m/s, which is float32 vs float64 accumulation and is the more accurate value.
- moving_average returned NaN almost everywhere on a gappy series (min_periods)
- save_animation returned None, so the FileLink cell could not work
- 04 read a peak-table column that the merged spectra module renamed

Verified: 02A 4/4 cells, 03A 48/48, 03B 82/82 (10 datasets), 04 13/13, 03T 7/7.
04's diurnal fits reproduce the pre-refactor diurnal_fits.csv.

+++ src/post_processing.py
@@ -102,7 +102,7 @@ def match_coefficients_to_times(
 
 def _broadcast(c: np.ndarray) -> np.ndarray:
     """(n_t,) -> (n_t, 1, 1) for broadcasting against (n_t, ny, nx) cubes."""
-    return np.asarray(c, dtype=float)[:, None, None]
+    return np.asarray(c, dtype=np.float32)[:, None, None]
 
 
 def apply_cubic(x: np.ndarray, c0: np.ndarray, c1: np.ndarray, c2: np.ndarray, c3: np.ndarray) -> np.ndarray:
